Renamed the components and containing folder
